Renamed: component to WatchedProducts.vue
