+++ apps/frontend/src/components/WatchedProducts.vue
